+++ src/main/java/com/telepathicgrunt/world_blender/mixin/MobSpawnInfoAccessor.java
@@ -24,5 +24,4 @@ public interface MobSpawnInfoAccessor {
 
     @Accessor("spawnCosts")
     void wb_setSpawnCosts(Map<EntityType<?>, MobSpawnInfo.SpawnCosts>  structureFeatures);
-
 }